Update deposits.json
I propose pulling Classref (name and value) and Entity (name and value). These fields were included in a prior version of several of my Quickbooks integrations running on Stitch and I do not see them in the latest version. These are key fields for my internal software. If agreeable, please update as soon as possible so that these fields will flow through in Stitch. Many thanks.